cpu-o3: Do not set Executed on load instruction to be replayed
A load instruction can be replayed if
1) it's strictly ordered or
2) it falls into load-store forwarding mismatch.
Case 1 was considered in executeLoad function but the case 2 wasn't.
It causes the case-2 replayed load instruction 2 to violate the assertion
condition "assert(!load_inst->isExecuted())" in LSQUnit::read.
This commit fixes the problem by adding consideration of the case 2 in
LSQUnit::executeLoad.

Change-Id: I2ebb6870313bf1c218709f251fd07e11bb1bb7ae