cpu: InstShiftAmt in TAGE-SC-L incorrectly used
The `instShiftAmt` in the branch prediction is intended for fixed
size instruction sets (Arm,RISC-V) to shift the FULL PC by 2 bits
as these two bits are useless to make a prediction (always zero).
TAGE-SC-L incorrectly used instShiftAmt to in the hash function
to xor a shifted version of the PC: pc ^ (pc >> instShiftAmt)
This can lead to the problem if the shift amount is set to 0 (in
case of x86) as will result in all bits 0 in for the PC.
Additionally, this commit adds the instShiftAmt to the statistical
corrector which did not use it.

Signed-off-by: David Schall <david.schall@tum.de>