mem: Fix bug in SysBridge + Xbar
Currently, if you use a system bridge immediately after a coherent
crossbar, when the packet returns the crossbar throws a panic that it
can't find the request.

The problem is that the SysBridge *changes* the pkt->req pointer to a
different pointer, and at the same time the XBar uses the pkt->req
pointer as the *key* in the routing table *after* calling
. Since it uses the pointer *after* , the
routing table uses the *new* (wrong) request pointer.

The solution in this change is to not modify the request pointer (which
was error prone anyway) and instead just change the RequestorID. I
believe there was not an API to change that when the SysBridge was
originally committed, but that API exists now.