arch-x86: Remove decodePages decode cache
The `decodePage/addrCacheMap` datastructure was meant as a *software*
cache to improve decode performance. The main problem is that the
`addrCacheMap` structure would grow without bounds as different
applications were run in the guest system.

To solve this problem, this change removes the cache. The result is a
more than 10x memory capacity savings for the simulator when booting
Ubuntu 24.04 and no performance change.

More details because the original code was difficult to understand:
- A new cache object (mapping physical addresses to data) was created
  *every time* the m5reg was updated with a unique value. This meant
  that there was a *new* cache created frequently.
- A new entry was created in the cache for every new address accessed.
  In other words, the "decodePages" cache (of which there was a copy for
  each unique value of the m5reg) would cache *all physical memory*
  which was accessed during fetch.

When running Ubuntu 24.04 boot with massif, the `addrCacheMap` was over
95% of all memory allocated by gem5. This was *not* a memory *leak*
because the data was always referenced in the map, but it was not a good
use of this memory. The memory footprint of this workload was over 18
GiB. (As a side note, this is why our daily tests were failing. They
were running out of memory.)

After this change, the time to complete the boot is basically unchanged
(maybe slightly faster), and the memory footprint is now under 2 GiB.

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>