mem-ruby: Add GPUCoalescer MSHR stats
This commit adds stats to the GPUCoalescer to track mshr misses and
hits-under-misses. The TCP hit counter is also incremented for mshr
hits-under-misses as MSHRs at the coalescer is equivalent to using TBEs
in the TCP Ruby protocol. Since coalescer MSHR misses go down to the
TCP, the miss counters are not incremented here