dev: rework PCI to add type1 header (#1949)
This adds the possiblity to have PCI device with type 1 config header
(PCI-to-PCI Bridge).

To achieve this, two classes are added `PciEndpoint` and `PciBridge`
which inherit the current `PciDevice`:

- `PciDevice` acts now on the common part of the configuration (based on
PCIe 5.0 specification, the first 16 bytes + capability pointer +
interrupt line/pin). This should not be inherited directly.
- `PciEndpoint` replace `PciDevice` for all devices connected as an
endpoint (GPU, network card, ...) and use the type 0 config header.
- `PciBridge` add the possibility for PCI bridges using the type 1
config header.

I tried to stick to the current architecture to avoid big breaking
changes. I'm open to any suggestion !

I also updated the `PXCAP` union for PCIe capabilities, which had a too
big offset before the device capabilities 2 register (20 instead of 16)
and also missing fields.

This is a first step into the bigger #1862 implementation.