arch-riscv: fix narrowing instructions with pin µop (#2026)
This patch fixes vector narrowing instructions, which where losing data
between µops to the same destination when a tail/mask agnostic policy
was being applied (i.e. no copy of old destination is done and
overwriting with 1s). The fix adds pinning to them so there's no loss of
data.