python: Remove m5.internal.params from SimObject
The m5.internal.params module just exposed some objects from _m5.
This change directly uses _m5.

Note that there was an unintended behavior in `getCCClass` where when
it was executed outside of the gem5 context (e.g., via gem5py) it would
throw an AttributeError that was silently caught. This change makes
the AttributeError happen both when the object doesn't have the attribute
and when _m5 isn't available. This keeps the prior behavior.

Signed-off-by: Jason Lowe-Power <jlowepower@google.com>