tests: Add cross product processor switching tests
This commit adds tests for processor switching between
combinations of atomic, timing, minor, and o3 cpus on the x86,
arm, and riscv ISAs. The SE mode tests use the matrix
multiplication workloads and run tests on one core, and
the FS mode tests use the NPB CG size S workload and run tests
on 4, 8, and 16 cores.