configs,gpu-compute,dev: MI355X and ROCm 7 support
This adds the CDNA4 ISA to the GfxVersion enum and handles conditionals
in configs, gpu-compute, and dev-amdgpu which enable certain features
for specific GFX versions. This is mostly the same as gfx942.

A new disk image with ROCm 7.0+ installed is required to use the MI355X
model. A preview disk image can be found at:

https://github.com/abmerop/gem5-resources/tree/rocm70

To use the model you can start with
configs/example/gem5_library/x86-mi300x-gpu.py and replace MI300X with
MI355X.

Note that most new instructions/features are available with the
exception of the new MFMA instructions.