misc: Add ExecContext to the tracer
With this commit we amend the InstTracer interface to accept a pointer
to the ExecContext. The key idea behind this is to allow a tracer to
extract dynamic information about the instruction stream (which is
not visible from the StaticInst class, and is instead available as an
example in the minor and o3 DynInst).

To avoid breaking compatibility for the time being, we overload
the getInstRecord and we keep supporting the version without
the ExecContext.

Change-Id: I23776f2d048e67432073397112e952923ec99037
Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>