dev: reworks PCI to add a PCI host bridge (#2298)
Due to mishandling on our original fork, which led us to delete it to
start over a clean one, this PR is a repost of #2204. Sorry for the
inconvenience.

This commit reworks the PCI architecture to:

  - Add a PCI host bridge
  - Add the possibility to have different upstream between multiple
    `PciDevice` (for future PCI-PCI bridge implementation)
  - Moves the responsibility to response a PCI configuration request in
    `PciDevice` instead of `PciHost`.

For that, the `PciUpstream` class is added and works like the old
`PciHost`, which now inherit from it. `PciDevice` will now need to
register to a `PciUpstream` instead of `PciHost` to get an instance of
`DeviceInterface`, this allows for the future implementation of a
PCI-to-PCI bridge. Also, bus number is now given to the device by its
upstream and not by the gem5 configuration which allow it to be changed
by the OS on boot for PCI-PCI bridge.

`PciUpstream` implementation has two member of the new class
`PciOneWayBridge` to implement the two way bridge two buses (e.g. system
and PCI bus). That class is a bridge with dynamic address ranges
depending on the ones from the connected buses.

The python `PciBus` class is also added, which simply inherit
from `NoncoherentXBar` and set the default port to an `IsaFake` to
answer to configuration request where no device respond.

So, the new PCI architecture is composed of a `PciHost`, connected to a
`PciBus` on PCI side to which all `PciDevice` are connected.

PCI configuration request are now handled by each `PciDevice` instead
of `PciHost` to mimic real PCI configuration.

This is another big step related to #1862.

It has an impact on existing gem5 configuration as now all PCI devices
need to be connected behind the `PciHost`. All configurations on the
repo have been updated to account that.
However, I'm not sure to understand correctly systems using Ruby and how
to connect on it, like with the GPU. For now, I connected PCI device's
indirectly through the bridge to Ruby, which works fine.