cpu-minor: modify numVecRegWrites and numVecPredRegWrites
This commit modifies how the execute stats numVecRegWrites and
numVecPredRegWrites are incremented in the Minor CPU. The changes
are taken from the implementation of the a similar function
(getWritableReg()) in the O3 CPU.
Prior to this change, the numVecRegWrites and
numVecPredRegWrites stats would not appear in the stats files
of simulations that used the Minor CPUs. After this change, they
now appear.