python,dev,arch-arm,arch-x86: Move Pins and Resets to params
The IntSource/SinkPin and ResetRequest/ResponsPorts were not SimObjects.
This causes some confusion (for humans) who expect these Port-like objects
to be in the same place as Ports, and it causes confusion for machines
becuase they are *not* SimObjects so they shouldn't be in m5.objects.
I don't fully understand the code path that allowed them to show up in
m5.objects, but I guess we were getting lucky with `from ... import *`
statements. As part of my work in breaking circular dependencies and
cleaning up python dependencies, this change moves the Pin and Reset
port-like objects into the m5/params package.

This requires updating some SimObjects, but this shouldn't be a user-facing
change.

Signed-off-by: Jason Lowe-Power <jlowepower@google.com>