stdlib: Allow more than 3GiB of CPU memory in VIPER board
More than 3GiB is a requirement for some GPU workloads, forcing users to
use the old configs. This change adds an additional "low memory" region
as well as a corresponding directory for that memory in the CPU cache
hierarchy. These changes are localized to VIPER board only as it
requires changing the CPU cache hierarchy specific to VIPER.