arch-arm,sim-se: Implement sigreturn for Arm64 (#2741)
The sigreturn system call takes CPU context data from the (usermode)
stack and applies it to the processor. This is usually used for
returning to the original program location after a signal, but it can
also be used to restore arbitrary program state (e.g., for SE-mode
"checkpoints").

This patch implements a subset of the system call to enable usermode
checkpoint restore using sigreturn.

Currently missing: Any floating point data

---------

Signed-off-by: Jason Lowe-Power <jlowepower@google.com>