misc: Add a PyPort to write to physmem from python (#2852)
The PyPort is the pybind11 interface around the PortProxy class. It is
meant to be used to read and write binary blobs to the guest memory from
python.

As an example, a PyPort can be returned from the System:

port = system.physPort

Its main APIs are the python based methods defined in
src/python/pybind11/port.cc. Those are simply:

def read(addr, size) -> bytes
def write(addr, data) -> None

To write an arbitrary binary blob to memory: at the moment we have an
heterogeneous set of solutions to do so. One is through the extras
parameters:

extras = VectorParam.String([], "Additional object files to load")
extras_addrs = VectorParam.Addr(
    [], "Load addresses for additional object files"
)

Though for some obscure reason those are only defined for
KernelWorkloads. "Simpler" environments (e.g. a traffic generator
platform with no binary) would not be able to use this static interface.
It also requires a path to a file in the the host FS for things to work
and does not allow inline slamming of binary data (sure, a python
platform can always write a binary blob to /tmp/ disk but it is not
ideal)

Another alternative is via the AbstractMemory.image_file parameter. The
caveat (on top of the requiring a file as solution above) is that

a) there is no way of specifying the offset of the image_file b) I
cannot specify multiple image files at the same time, so the only
workaround to this is to define multiple memories

This python port solution fixes all aforementioned shortcomings and
therefore has the following perks:

1) It's a generall solution which works with every configuration

2) We can use it to dynamically patch (write) the content of memory
during simulation from python

3) We can use it to dynamicall read the content of memory during
simulation from python for debug purposes

Change-Id: I99b274fb64504d922044c0b8cf24e6ad75060886

Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Co-authored-by: Giacomo Travaglini <giacomo.travaglini.arm.com>